(fix)correcao do tipo do repositorio, estava gerando erro no swagger

+++ src/main/java/br/com/duxusdesafio/repository/IntegranteRepository.java
@@ -3,5 +3,5 @@ package br.com.duxusdesafio.repository;
 import br.com.duxusdesafio.model.Integrante;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IntegranteRepository extends JpaRepository<Integrante, Integer> {
+public interface IntegranteRepository extends JpaRepository<Integrante,Long> {
 }
